Split playes service and player repository
