+++ W3ChampionsStatisticService/Services/PlayerStatsService.cs
@@ -16,7 +16,9 @@ namespace W3ChampionsStatisticService.Services
         private readonly ICachedDataProvider<SeasonMapInformation> _seasonMapCached;
         private readonly IW3StatsRepo _w3StatsRepo;
 
-        public PlayerStatisticsService(IPlayerStatsRepository playerStatsRepository, MatchmakingProvider matchmakingProvider, ICachedDataProvider<SeasonMapInformation> seasonMapCached, IW3StatsRepo w3StatsRepo)
+        public PlayerStatisticsService(IPlayerStatsRepository playerStatsRepository,
+            MatchmakingProvider matchmakingProvider, ICachedDataProvider<SeasonMapInformation> seasonMapCached,
+            IW3StatsRepo w3StatsRepo)
         {
             _playerStatsRepository = playerStatsRepository;
             _matchmakingProvider = matchmakingProvider;
@@ -64,5 +66,6 @@ namespace W3ChampionsStatisticService.Services
 
             return loadMatchesOnMap;
         }
+
     }
 }
