+++ WC3ChampionsStatisticService.UnitTests/Performance/DBPerformanceTest.cs
@@ -132,7 +132,7 @@ namespace WC3ChampionsStatisticService.Tests.Matches
         {
             Stopwatch sw = new Stopwatch();
             sw.Start();
-            var playerRepository = new PlayerRepository(MongoClient, null, CreateTestCache<List<MmrRank>>());
+            var playerRepository = new PlayerRepository(MongoClient);
             var playerLoadedAgain = await playerRepository.LoadRaceStatPerGateway("ShaDeFaDe#2441", GateWay.Europe, 1);
             for (int i = 0; i < 1000; i++)
             {
@@ -146,7 +146,7 @@ namespace WC3ChampionsStatisticService.Tests.Matches
         {
             Stopwatch sw = new Stopwatch();
             sw.Start();
-            var playerRepository = new PlayerRepository(MongoClient, null, CreateTestCache<List<MmrRank>>());
+            var playerRepository = new PlayerRepository(MongoClient);
             var playerLoadedAgain = await playerRepository.LoadGameModeStatPerGateway("ShaDeFaDe#2441", GateWay.Europe, 1);
             for (int i = 0; i < 1000; i++)
             {
