+++ WC3ChampionsStatisticService.UnitTests/Player/PlayerTests.cs
@@ -10,6 +10,7 @@ using W3ChampionsStatisticService.PlayerProfiles.GameModeStats;
 using W3ChampionsStatisticService.PlayerProfiles.MmrRankingStats;
 using W3ChampionsStatisticService.PlayerProfiles.RaceStats;
 using W3C.Contracts.Matchmaking;
+using W3ChampionsStatisticService.Services;
 
 namespace WC3ChampionsStatisticService.Tests.Player
 {
@@ -19,7 +20,7 @@ namespace WC3ChampionsStatisticService.Tests.Player
         [Test]
         public async Task LoadAndSave()
         {
-            var playerRepository = new PlayerRepository(MongoClient, null, CreateTestCache<List<MmrRank>>());
+            var playerRepository = new PlayerRepository(MongoClient);
 
             var player = PlayerOverallStats.Create("peter#123");
             await playerRepository.UpsertPlayer(player);
@@ -31,7 +32,7 @@ namespace WC3ChampionsStatisticService.Tests.Player
         [Test]
         public async Task Search()
         {
-            var playerRepository = new PlayerRepository(MongoClient, null, CreateTestCache<List<MmrRank>>());
+            var playerRepository = new PlayerRepository(MongoClient);
 
             var player = PlayerOverallStats.Create("Peter#123");
             var player2 = PlayerOverallStats.Create("Anderer#123");
@@ -46,7 +47,7 @@ namespace WC3ChampionsStatisticService.Tests.Player
         [Test]
         public async Task Search_DuplicateNameBug()
         {
-            var playerRepository = new PlayerRepository(MongoClient, null, CreateTestCache<List<MmrRank>>());
+            var playerRepository = new PlayerRepository(MongoClient);
 
             var player = PlayerOverallStats.Create("ThunderHorn#2481");
             var player2 = PlayerOverallStats.Create("ThunderHorn#21132");
@@ -62,7 +63,7 @@ namespace WC3ChampionsStatisticService.Tests.Player
         [Test]
         public async Task Search_DuplicateNameBug_RefineWithBtag()
         {
-            var playerRepository = new PlayerRepository(MongoClient, null, CreateTestCache<List<MmrRank>>());
+            var playerRepository = new PlayerRepository(MongoClient);
 
             var player = PlayerOverallStats.Create("ThunderHorn#2481");
             var player2 = PlayerOverallStats.Create("ThunderHorn#21132");
@@ -77,8 +78,9 @@ namespace WC3ChampionsStatisticService.Tests.Player
         [Test]
         public async Task GlobalSearch()
         {
-            var playerRepository = new PlayerRepository(MongoClient, personalSettingsProvider, CreateTestCache<List<MmrRank>>());
+            var playerRepository = new PlayerRepository(MongoClient);
             var personalSettingsRepository = new PersonalSettingsRepository(MongoClient);
+            var playerService = new PlayerService(playerRepository, CreateTestCache<List<MmrRank>>(), personalSettingsProvider);
 
             var player1 = new PersonalSetting("ThunderHorn#2481");
             var playerStats = PlayerOverallStats.Create("ThunderHorn#2481");
@@ -99,12 +101,12 @@ namespace WC3ChampionsStatisticService.Tests.Player
             var player3 = new PersonalSetting("OtherPlayer#123");
             await personalSettingsRepository.Save(player3);
 
-            var players = await playerRepository.GlobalSearchForPlayer("under");
+            var players = await playerService.GlobalSearchForPlayer("under");
             Assert.AreEqual(2, players.Count);
             Assert.AreEqual(player2.Id, players[0].BattleTag);
             Assert.AreEqual(player1.Id, players[1].BattleTag);
 
-            players = await playerRepository.GlobalSearchForPlayer("under", "ThunderHorn#21132");
+            players = await playerService.GlobalSearchForPlayer("under", "ThunderHorn#21132");
             Assert.AreEqual(1, players.Count);
             Assert.AreEqual(player1.Id, players[0].BattleTag);
             Assert.AreEqual(player1.ProfilePicture.PictureId, players[0].ProfilePicture.PictureId);
@@ -114,7 +116,7 @@ namespace WC3ChampionsStatisticService.Tests.Player
         [Test]
         public async Task PlayerMapping()
         {
-            var playerRepository = new PlayerRepository(MongoClient, null, CreateTestCache<List<MmrRank>>());
+            var playerRepository = new PlayerRepository(MongoClient);
 
             var player = PlayerOverallStats.Create("peter#123");
             player.RecordWin(Race.HU, 0, true);
@@ -132,7 +134,7 @@ namespace WC3ChampionsStatisticService.Tests.Player
         [Test]
         public async Task PlayerStatsMapping()
         {
-            var playerRepository = new PlayerRepository(MongoClient, null, CreateTestCache<List<MmrRank>>());
+            var playerRepository = new PlayerRepository(MongoClient);
 
             var battleTagIdCombined = new BattleTagIdCombined(new List<PlayerId>
                 {
@@ -156,7 +158,7 @@ namespace WC3ChampionsStatisticService.Tests.Player
         [Test]
         public async Task PlayerIdMappedRight()
         {
-            var playerRepository = new PlayerRepository(MongoClient, null, CreateTestCache<List<MmrRank>>());
+            var playerRepository = new PlayerRepository(MongoClient);
 
             var player1 = PlayerOverallStats.Create("peter#123");
             var player2 = PlayerOverallStats.Create("wolf#456");
@@ -173,7 +175,7 @@ namespace WC3ChampionsStatisticService.Tests.Player
         [Test]
         public async Task PlayerMultipleWinRecords()
         {
-            var playerRepository = new PlayerRepository(MongoClient, null, CreateTestCache<List<MmrRank>>());
+            var playerRepository = new PlayerRepository(MongoClient);
             var personalSettingsRepo = new PersonalSettingsRepository(MongoClient);
 
             var handler = new PlayerOverallStatsHandler(playerRepository, personalSettingsRepo);
@@ -207,7 +209,7 @@ namespace WC3ChampionsStatisticService.Tests.Player
         [Test]
         public async Task PlayersFromAn1v1Match_GameModeHandler_Test()
         {
-            var playerRepository = new PlayerRepository(MongoClient, null, CreateTestCache<List<MmrRank>>());
+            var playerRepository = new PlayerRepository(MongoClient);
             var handler = new PlayerGameModeStatPerGatewayHandler(playerRepository);
 
             var ev = TestDtoHelper.CreateFakeEvent();
@@ -233,7 +235,7 @@ namespace WC3ChampionsStatisticService.Tests.Player
         [Test]
         public async Task PlayersFromAn2v2ATMatch_GameModeHandler_Test()
         {
-            var playerRepository = new PlayerRepository(MongoClient, null, CreateTestCache<List<MmrRank>>());
+            var playerRepository = new PlayerRepository(MongoClient);
             var handler = new PlayerGameModeStatPerGatewayHandler(playerRepository);
 
             var ev = TestDtoHelper.CreateFake2v2AtEvent();
@@ -266,7 +268,7 @@ namespace WC3ChampionsStatisticService.Tests.Player
         [Test]
         public async Task PlayersFromAn2v2AT_vs_RTMatch_GameModeHandler_Test()
         {
-            var playerRepository = new PlayerRepository(MongoClient, null, CreateTestCache<List<MmrRank>>());
+            var playerRepository = new PlayerRepository(MongoClient);
             var handler = new PlayerGameModeStatPerGatewayHandler(playerRepository);
 
             var ev = TestDtoHelper.CreateFake2v2AtEvent();
@@ -305,7 +307,7 @@ namespace WC3ChampionsStatisticService.Tests.Player
         [Test]
         public async Task PlayersFromAnFFAMatch_GameModeHandler_Test()
         {
-            var playerRepository = new PlayerRepository(MongoClient, null, CreateTestCache<List<MmrRank>>());
+            var playerRepository = new PlayerRepository(MongoClient);
             var handler = new PlayerGameModeStatPerGatewayHandler(playerRepository);
 
             var ev = TestDtoHelper.CreateFakeFFAEvent();
@@ -347,7 +349,7 @@ namespace WC3ChampionsStatisticService.Tests.Player
         {
             var matchFinishedEvent = TestDtoHelper.CreateFakeFootmenFrenzyEvent();
 
-            var playerRepository = new PlayerRepository(MongoClient, null, CreateTestCache<List<MmrRank>>());
+            var playerRepository = new PlayerRepository(MongoClient);
             var handler = new PlayerGameModeStatPerGatewayHandler(playerRepository);
 
             await handler.Update(matchFinishedEvent);
@@ -380,7 +382,7 @@ namespace WC3ChampionsStatisticService.Tests.Player
         [Test]
         public async Task Player_UpdateMmrRpTimeline()
         {
-            var playerRepository = new PlayerRepository(MongoClient, null, CreateTestCache<List<MmrRank>>());
+            var playerRepository = new PlayerRepository(MongoClient);
             var handler = new PlayerMmrRpTimelineHandler(playerRepository);
             var matchFinishedEvent1 = TestDtoHelper.CreateFakeEvent();
             var matchFinishedEvent2 = TestDtoHelper.CreateFakeEvent();
